// src/app/api/model/bottleType.ts
/**
 * Cave à vins
 * API cave à vins
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */



export interface BottleType {
    id?: number;
    name?: string;
    date?: number;
    region?: string;
    colour?: string;
    valide?: boolean;
}